refactor: filename from project registry to projects
